Created are interface «Sortable» and abstract class «Sorter»; Corrected classes «SortingManager» and «BubbleSort»; Checked the health of the tests.
