+++ Sorting/Test/ru/svdegron/java2021/sort/BubbleSortTest.java
@@ -7,11 +7,11 @@ import org.junit.Test;
 
 public class BubbleSortTest {
 
-	private BubbleSort sorter;
+	private BubbleSort bubbleSort;
 	
 	@Before
 	public void setUp() {
-		sorter = new BubbleSort();
+		bubbleSort = new BubbleSort();
 	}
 
 	@Test
@@ -20,7 +20,7 @@ public class BubbleSortTest {
 		int[] unsorted = { 4, 3, 2, 1 };
 		int[] sorted = { 1, 2, 3, 4 };
 		
-		assertArrayEquals(sorted, sorter.AscendingFromBegin(unsorted));
+		assertArrayEquals(sorted, bubbleSort.AscendingFromBegin(unsorted));
 		
 	}
 	
@@ -30,7 +30,7 @@ public class BubbleSortTest {
 		int[] unsorted = { 8, 7, 6, 5 };
 		int[] sorted = { 5, 6, 7, 8 };
 		
-		assertArrayEquals(sorted, sorter.AscendingFromEnd(unsorted));
+		assertArrayEquals(sorted, bubbleSort.AscendingFromEnd(unsorted));
 		
 	}
