+++ Sorting/src/ru/svdegron/java2021/sort/Sortable.java
@@ -1,0 +1,9 @@
+package ru.svdegron.java2021.sort;
+
+public interface Sortable {
+	
+	// this is methods most interesting, for my opinion
+	public int[] AscendingFromBegin(int[] input);
+	public int[] DescendingFromEnd(int[] input);
+
+}
